Configure network-manager on VM
anyconnect appears to detect the presence of dhclient when setting up
the routes, and leaves the route pointing at the DHCP server routed
directly as opposed to via the VPN. Anyconnect doesn't appear to detect
when systemd-networkd is handling the dhcp client, and therefore doesn't
configure the routes as necessary. This has the effect of preventing ssh
connections to the guest from working after the VPN is connected.